+++ examples/graphql/resolvers.php
@@ -1,50 +1,51 @@
-<?php
+<?php declare(strict_types=1);
 
-use RedBeanPHP\R;
+namespace Siler\Example\GraphQL;
+
+use DateTime;
+use RedBeanPHP\{OODBBean, R};
 use Siler\GraphQL;
-use function Siler\array_get_arr;
+use function Siler\{array_get_arr, array_get_str};
 
 R::setup('sqlite:' . __DIR__ . '/db.sqlite');
 
-$roomByName = function ($name) {
+$room_by_name = function ($name): ?OODBBean {
     return R::findOne('room', 'name = ?', [$name]);
 };
 
-$roomType = [
+$room = [
     'messages' => function ($room) {
         return R::findAll('message', 'room_id = ?', [$room['id']]);
     }
 ];
 
-$queryType = [
+$queries = [
     'rooms' => function () {
         return R::findAll('room');
     },
-    'messages' => function ($root, $args) use ($roomByName) {
-        $roomName = $args['roomName'];
-        $room = $roomByName($roomName);
-        $messages = R::find('message', 'room_id = ?', [$room['id']]);
-
-        return $messages;
+    'messages' => function ($_, $args) use ($room_by_name) {
+        $room_name = array_get_str($args, 'roomName');
+        $room = $room_by_name($room_name);
+        return R::find('message', 'room_id = ?', [$room['id']]);
     }
 ];
 
-$mutationType = [
-    'start' => function ($root, $args) {
-        $roomName = $args['roomName'];
+$mutations = [
+    'start' => function ($_, $args) {
+        $room_name = array_get_str($args, 'roomName');
 
         $room = R::dispense('room');
-        $room['name'] = $roomName;
+        $room['name'] = $room_name;
 
         R::store($room);
 
         return $room;
     },
-    'chat' => function ($root, $args) use ($roomByName) {
-        $roomName = $args['roomName'];
-        $body = $args['body'];
+    'chat' => function ($_, $args) use ($room_by_name) {
+        $room_name = array_get_str($args, 'roomName');
+        $body = array_get_str($args, 'body');
 
-        $room = $roomByName($roomName);
+        $room = $room_by_name($room_name);
 
         $message = R::dispense('message');
         $message['roomId'] = $room['id'];
@@ -53,29 +54,39 @@ $mutationType = [
 
         R::store($message);
 
-        $message['roomName'] = $roomName;
+        $message['roomName'] = $room_name;
         // For the inbox filter
         GraphQL\publish('inbox', $message);
         // <- Exactly what "inbox" will receive
         return $message;
     },
-    'upload' => function ($root, $args) {
+    'close' => function ($_, array $args) use ($room_by_name): bool {
+        $room_name = array_get_str($args, 'roomName');
+        $room = $room_by_name($room_name);
+
+        if ($room === null) {
+            return false;
+        }
+
+        R::trash($room);
+        return true;
+    },
+    'upload' => function ($_, $args) {
         $file = array_get_arr($args, 'file');
         move_uploaded_file($file['tmp_name'], __DIR__ . "/uploads/{$file['name']}");
         return ['name' => $file['name']];
     },
 ];
 
-$subscriptionType = [
+$subscriptions = [
     'inbox' => function ($message) {
-        // <- Received from "publish"
         return $message;
     }
 ];
 
 return [
-    'Room' => $roomType,
-    'Query' => $queryType,
-    'Mutation' => $mutationType,
-    'Subscription' => $subscriptionType
+    'Room' => $room,
+    'Query' => $queries,
+    'Mutation' => $mutations,
+    'Subscription' => $subscriptions,
 ];